+++ phi_cloud_action/data/example/archives.py
@@ -2,14 +2,13 @@
 # ----------------------- 导包区喵 -----------------------
 from sys import argv
 
-from ..PhiCloudLib import (
+from phi_cloud_action import (
     PhigrosCloud,
     readDifficultyFile,
     decryptSave,
     unzipSave,
     formatSaveDict,
-    getB19,
-    logger,
+    checkSaveHistory,
 )
 
 
@@ -26,14 +25,14 @@ else:
 # ----------------------- 运行区喵 -----------------------
 
 
-def printB19(token):
+def archives(token):
     with PhigrosCloud(token) as cloud:
 
         # 读取难度定数文件喵
-        difficulty = readDifficultyFile()
-
+        difficulty = readDifficultyFile() 
+        
         # 获取玩家summary喵
-        summary = cloud.getSummary() 
+        summary = cloud.getSummary()  
 
         # 获取并解析存档喵
         save_data = cloud.getSave(summary["url"], summary["checksum"])
@@ -41,17 +40,8 @@ def printB19(token):
         save_dict = decryptSave(save_dict)
         save_dict = formatSaveDict(save_dict)
 
-        # 获取b19喵
-        b19 = getB19(save_dict["gameRecord"], difficulty)
-
-        # 输出并计算玩家rks喵
-        rks = 0.0
-        logger.info("玩家b19：")
-        for song in b19:
-            logger.info(song)
-            rks += song["rks"]
-        logger.info(f"玩家rks：{rks / 20}")
-
+        # 存储存档历史记录喵
+        checkSaveHistory(token, summary, save_data, difficulty)
 
 if __name__ == "__main__":
-    printB19(sessionToken)
+    archives(sessionToken)
