+++ phi_cloud_action/data/example/printNickname.py
@@ -2,7 +2,7 @@
 # ----------------------- 导包区喵 -----------------------
 from sys import argv
 
-from ..PhiCloudLib import (
+from phi_cloud_action import (
     PhigrosCloud,
     logger,
 )
