--- phi_cloud_action/example/__init__.py
@@ -1,6 +1,0 @@
-from .archives import archives
-from .getSave import getSave
-from .printB19 import printB19
-from .printNickname import printNickname
-from .uploadNickname import uploadNickname
-from .other import (refresh_token)
